+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.14] - 2025-11-18
+
+### Added
+- SVG Graph Network library integration for blog post demos
+- Demo graph data files for command tree and library features visualization
+
 ## [0.10.13] - 2025-11-16
 
 ### Fixed